tests/wpt/tests/trusted-types/Element-setAttributeNS.html

<!DOCTYPE html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="support/helper.sub.js"></script>
  <script src="support/namespaces.js"></script>
</head>
<body>
<script>
    test(t => {
      assert_element_accepts_trusted_html_set_ns(window, '0', t, 'a', 'b', RESULTS.HTML);
    }, "Element.setAttributeNS assigned via policy (successful HTML transformation)");

    test(t => {
      assert_element_accepts_trusted_script_set_ns(window, '1', t, 'a', 'b', RESULTS.SCRIPT);
    }, "Element.setAttributeNS assigned via policy (successful Script transformation)");

    test(t => {
      assert_element_accepts_trusted_script_url_set_ns(window, '2', t, 'a', 'b', RESULTS.SCRIPTURL);
    }, "Element.setAttributeNS assigned via policy (successful ScriptURL transformation)");

    test(t=> {
      const policy = trustedTypes.createPolicy("p", { createScriptURL: s => s + s });
      assert_element_accepts_value_set_ns("image", "href", policy.createScriptURL("v"), "vv",
        NSURI_SVG, NSURI_XLINK);
    }, "Element.setAttributeNS accepts a URL on <svg:image xlink:href/>");
</script>
